% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ascale}
\alias{ascale}
\title{Arbitrary scaling}
\usage{
ascale(x, .mean, .sd)
}
\arguments{
\item{x}{a numeric vector}

\item{.mean}{the mean of the output}

\item{.sd}{the sd of the output}
}
\value{
a rescaled numeric vector
}
\description{
Rescale a vector to a given mean and variance
}

