% Generated by roxygen2 (4.0.1): do not edit by hand
\name{\%rbyname\%}
\alias{\%rbyname\%}
\title{Create list element from objects in supplied list and give them
the names of the objects as element names}
\usage{
structure \%rbyname\% x
}
\arguments{
\item{structure}{a list object to append to}

\item{x}{a list holding the objects to be added to the list}
}
\value{
nothing, invoked for side effects: the original list is
modified in place
}
\description{
Instead of mylist$mysomething <- mysomething; mylist$somethingelse
<- somethingelse, we can now write mylist %byname%
list(mysomething, somethingelse). A little easier.
}

